refactor: RestTemplateConfig.java 패키지 변경
